Use ints instead of longs for LZSS stuff
Apparently C# (or at least the older versions) have a 2GB object size
limit, so arrays like the file buffer will never surpass int's limit
anyway.

Also trying to address the performance issue the compressors are having
right now.

+++ Comper/Comper.impl.cs
@@ -6,25 +6,25 @@
 
     public static partial class Comper
     {
-        private static void FindExtraMatches(ushort[] data, long pos, long data_size, long offset, LZSS.NodeMeta[] node_meta_array)
+        private static void FindExtraMatches(ushort[] data, int pos, int data_size, int offset, LZSS.NodeMeta[] node_meta_array)
         {
             // Comper has no special matches
         }
 
-        private static long GetMatchCost(long distance, long length)
+        private static int GetMatchCost(int distance, int length)
         {
             return 1 + 16;
         }
 
         internal static void Encode(Stream source, Stream destination)
         {
-            long size_bytes = source.Length - source.Position;
+			int size_bytes = (int)(source.Length - source.Position);
             byte[] buffer_bytes = new byte[size_bytes + (size_bytes & 1)];
-            source.Read(buffer_bytes, 0, (int)size_bytes);
+            source.Read(buffer_bytes, 0, size_bytes);
 
-            long size = (size_bytes + 1) / 2;
+			int size = (size_bytes + 1) / 2;
             ushort[] buffer = new ushort[size];
-            for(long i = 0; i < size; ++i)
+            for(int i = 0; i < size; ++i)
             {
                 buffer[i] = (ushort)((buffer_bytes[i * 2] << 8) | buffer_bytes[(i * 2) + 1]);
             }
@@ -34,12 +34,12 @@
             UInt16BE_NE_H_OutputBitStream bitStream = new UInt16BE_NE_H_OutputBitStream(destination);
             MemoryStream data = new MemoryStream();
 
-            for (long node_index = 0; node_meta_array[node_index].next_node_index != long.MaxValue; node_index = node_meta_array[node_index].next_node_index)
+            for (int node_index = 0; node_meta_array[node_index].next_node_index != int.MaxValue; node_index = node_meta_array[node_index].next_node_index)
             {
-                long next_index = node_meta_array[node_index].next_node_index;
+				int next_index = node_meta_array[node_index].next_node_index;
 
-                long length = node_meta_array[next_index].match_length;
-                long distance = next_index - node_meta_array[next_index].match_length - node_meta_array[next_index].match_offset;
+				int length = node_meta_array[next_index].match_length;
+				int distance = next_index - node_meta_array[next_index].match_length - node_meta_array[next_index].match_offset;
 
                 if (length != 0)
                 {
@@ -104,7 +104,7 @@
                         break;
                     }
 
-                    for (long i = 0; i <= length; i++)
+                    for (int i = 0; i <= length; i++)
                     {
                         long writePosition = destination.Position;
                         destination.Seek(writePosition - distance, SeekOrigin.Begin);
